fix lint errors

+++ contracts/src/AccountRegistry.sol
@@ -468,15 +468,14 @@ contract AccountRegistry is Initializable, EIP712Upgradeable, Ownable2StepUpgrad
         uint256 leafIndex = nextLeafIndex;
 
         uint256 bitmap = 0;
-        for (uint256 i = 0; i < authenticatorAddresses.length; i++) {
+        for (uint32 i = 0; i < authenticatorAddresses.length; i++) {
             address authenticatorAddress = authenticatorAddresses[i];
             if (authenticatorAddress == address(0)) {
                 revert ZeroAddress();
             }
 
             _validateNewAuthenticatorAddress(authenticatorAddress);
-            authenticatorAddressToPackedAccountData[authenticatorAddress] =
-                PackedAccountData.pack(leafIndex, 0, uint32(i));
+            authenticatorAddressToPackedAccountData[authenticatorAddress] = PackedAccountData.pack(leafIndex, 0, i);
             bitmap = bitmap | (1 << i);
         }
         _setRecoveryAddressAndBitmap(leafIndex, recoveryAddress, bitmap);
